Added "Joker" (null), changed returnCardString
Cards that are named Joker or have a Joker rank are non-cards for the purposes of the game. They are necessary to be able to inform the program that the deck is out of cards without breaking everything. Add a check for Joker every time you call a new card from the deck,