Finalized Javadoc, added more methods, cleaned up code
Deck now has a built-in check for when the deck is empty. If you attempt to call a card when the deck is empty via drawCard or drawRandomCard, it will return a Joker card. The Joker card is not to be used in the game, so the game must implement a check for it.

Changed the way the constructors work to streamline the process. If you initialize a Deck, it will NOT shuffle the newly generated deck unless you also specified maximumCards and random. So, if you do Deck deck = new Deck(), you must also do deck.shuffle() afterwards if you want to shuffle it.

You can only shuffle the deck once. This is due to the way it was implemented. I could change this behavior, but that would be a fundamental rewrite of all the code and for the purposes of the game, it is not necessary.

Javadoc is finalized. Most if not all of the usability of the deck class is documented there.

printDeck now prints a special message when the deck is empty.