Added image assignments, cleaned up code
Note: I can't test this, not properly at least, until JavaFX is up and running on this project. Because of that, right now it returns a buffered image; the final code might need something else, but I cannot verify exactly what's needed if the JavaFX part of this is not set up.