Because Deck was so good...
Deck2 is meant to be an alternative to Deck, it fully implements everything Deck has but it also does some more things, such as allowing you to shuffle the deck multiple times within a single instance, resetting the deck whenever you want to start again, and adding a "History", of sorts, with the Played Cards.

The reason why Deck2 is not replacing Deck is due to compatibility. It handles methods and constructors differently, which means that it is not seamless to switch between them. If you've already started using Deck, then switching for this game is not really necessary, but otherwise use Deck2.